Implement metadata rest for i64 and array
Sadly the rest of the types are inaccessible right now as I have to use
the Any trait, which requires 'static lifetime, which I don't have as
the values are borrowed from the metadata type.

Will try to improve the type as presented in the mpris library at some
point.